move common.EmailAddress out
